requirements.txt
Flask
requests
beautifulsoup4
python-dotenv
gunicorn
openai>=1.3.0
flask-cors
supabase
psycopg2-binary